requirements.txt
python-telegram-bot
langchain
langchain-openai
langchain-community
chromadb
pypdf
sentence-transformers
python-dotenv